wilber/requirements/base.txt
# Django
# ------------------------------------------------------------------------------
django==3.0.5  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-filter==2.0.0
django-allauth==0.41.0
django-autoslug==1.9.4
django-imagekit==4.0.2
django-rest-auth==0.9.5
django-extensions==2.1.6
django-request==1.5.4
django-mass-edit==3.2.0
django-bulma==0.6.0
django-webpack-loader==0.6.0
djangorestframework==3.11.0
djangorestframework-jwt

Markdown==3.0.1
Pillow==7.1.1
pytz==2019.3
sentry-sdk==0.11.2
